fix(dataset): support getting dataset info
GET requests cannot have a body, our API contained a bug where GET was trying to parse a nonexistent body, and were silently returning 200 status. The fix is to accept dataset ref strings on the /dataset/ endpoint, parse it, and use it for lookup

To do this properly, we need access to repo.DatasetRef's parse methods, which we can't have directly due to cyclical imports. I've always wanted to have Qri's naming system compose it's own package, but we have a lot of thinking to do around how a change like this would affect our codebase. For now I've stubbed out what a new ns package would look like and put it in registry for us to think about, knowing that it'll either move or be removed. At least this way we can ship a fix to registry, and we get the bonus of some code to consider